+++ .gitignore
@@ -2,6 +2,7 @@
 /dist
 /node_modules
 env
+/prisma/migrations
 
 # Logs
 logs
