+++ src/app/mobile-user/auth/use-cases/auth-sign-refresh-token.ts
@@ -14,7 +14,7 @@ export class SignRefreshTokenUseCase {
   async execute(
     phone: string,
     id: number,
-  ): Promise<{ token: string; expirationDate: string }> {
+  ): Promise<any> {
     const payload: TokenPayload = { phone: phone, clientId: id };
     const secret = this.configService.get<string>('jwtRefreshTokenSecret');
     const expiresIn = this.configService.get<string>(
