+++ src/app/mobile-user/otp/interfaces/otp.ts
@@ -4,5 +4,4 @@ export abstract class IOtpRepository {
   abstract create(input: Otp): Promise<Otp>;
   abstract findOne(phone: string): Promise<Otp>;
   abstract removeOne(phone: string): Promise<void>;
-  abstract send(input: Otp): Promise<any>;
 }
