+++ src/libs/sms/adapter.ts
@@ -1,0 +1,5 @@
+import { Otp } from '@mobile-user/otp/domain/otp';
+
+export abstract class ISmsAdapter {
+  abstract send(input: Otp, text: string): Promise<any>;
+}
